# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Private local photo library API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      summary: Get current local session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
    post:
      operationId: login
      tags: [auth]
      summary: Log in to the local library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Invalid credentials
    delete:
      operationId: logout
      tags: [auth]
      summary: End the current session
      responses:
        "204":
          description: Logged out
...
  /albums:
    get:
      operationId: listAlbums
      tags: [albums]
      summary: List albums
      responses:
        "200":
          description: Albums
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Album"
    post:
      operationId: createAlbum
      tags: [albums]
      summary: Create an album
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumInput"
      responses:
        "201":
          description: Created album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{albumId}:
    patch:
      operationId: updateAlbum
      tags: [albums]
      summary: Rename an album
      parameters:
        - $ref: "#/components/parameters/AlbumId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlbumUpdate"
      responses:
        "200":
          description: Updated album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
    delete:
      operationId: deleteAlbum
      tags: [albums]
      summary: Delete an album without deleting media
      parameters:
        - $ref: "#/components/parameters/AlbumId"
      responses:
        "204":
          description: Deleted
  /albums/{albumId}/photos:
    post:
      operationId: addPhotosToAlbum
      tags: [albums]
      summary: Add photos to an album
      parameters:
        - $ref: "#/components/parameters/AlbumId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoIdsInput"
      responses:
        "200":
          description: Updated album
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Album"
  /albums/{albumId}/photos/{photoId}:
    delete:
      operationId: removePhotoFromAlbum
      tags: [albums]
      summary: Remove a photo from an album
      parameters:
        - $ref: "#/components/parameters/AlbumId"
        - $ref: "#/components/parameters/PhotoId"
      responses:
        "204":
          description: Removed
...
components:
  parameters:
    PhotoId:
      in: path
      name: photoId
      required: true
      schema: { type: string }
    AlbumId:
      in: path
      name: albumId
      required: true
      schema: { type: string }
...
  schemas:
...
    Session:
      type: object
      properties:
        authenticated: { type: boolean }
        username: { type: ["string", "null"] }
      required: [authenticated, username]
    LoginInput:
      type: object
      properties:
        username: { type: string, minLength: 1 }
        password: { type: string, minLength: 1 }
      required: [username, password]
...
    Album:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        photoCount: { type: number }
        coverUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, photoCount, coverUrl, createdAt, updatedAt]
    AlbumInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
      required: [name]
    AlbumUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
      required: [name]
    PhotoIdsInput:
      type: object
      properties:
        photoIds:
          type: array
          items: { type: string }
          minItems: 1
      required: [photoIds]
...