# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Private local photo library API
 * OpenAPI spec version: 0.1.0
 */

export interface Album {
  id: string;
  name: string;
  photoCount: number;
  /** @nullable */
  coverUrl: string | null;
  createdAt: Date;
  updatedAt: Date;
}