# lib/api-zod/src/generated/types/albumInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Private local photo library API
 * OpenAPI spec version: 0.1.0
 */

export interface AlbumInput {
  /** @minLength 1 */
  name: string;
}